# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant access control and smart building platform API.\n\nAll endpoints are tenant-scoped through the `Authorization: Bearer dm3_live_...`\nor `dm3_test_...` API token. Issue tokens from Settings → API Integration.\nTokens prefixed `dm3_live_` act on production data; `dm3_test_` are sandbox.\n\nThe default rate limit is 60 requests/minute per token. Responses use a\nconsistent envelope: success responses return the resource directly, errors\nreturn `{ \"error\": \"\u003ccode\u003e\", \"message\": \"\u003chuman readable\u003e\" }`.",
        "title": "Duall Master API",
        "termsOfService": "https://duali.com/terms",
        "contact": {
            "name": "Duall Master Support",
            "url": "https://duali.com/support",
            "email": "support@duali.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/attendance/holidays": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Plugin-gated (attendance). Reads open to any attendance-enabled user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "List holidays",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by year (defaults to current year)",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_duali_dm3-backend_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/attendance/me/attendance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Plugin-gated (attendance). Returns the caller's own attendance records over the requested window.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendance"
                ],
                "summary": "My attendance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISO date (YYYY-MM-DD) start of window. Defaults to first day of current month.",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO date (YYYY-MM-DD) end of window. Defaults to today.",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_duali_dm3-backend_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_duali_dm3-backend_pkg_httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API token in the form `Bearer dm3_live_...` or `Bearer dm3_test_...`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}